add more cuStateVec support / refactor qubitvector_thrust and chunk_container
