+++ src/simulators/statevector/chunk/chunk_container.hpp
@@ -63,6 +63,8 @@ DISABLE_WARNING_POP
 #include "simulators/statevector/chunk/cuda_kernels.hpp"
 #endif
 
+#include "simulators/statevector/chunk/thrust_kernels.hpp"
+
 namespace AER {
 namespace QV {
 
@@ -77,391 +79,6 @@ struct BlockedGateParams
   unsigned char qubit_;
 };
 
-//========================================
-//  base class of gate functions
-//========================================
-template <typename data_t>
-class GateFuncBase
-{
-protected:
-  thrust::complex<data_t>* data_;   //pointer to state vector buffer
-  thrust::complex<double>* matrix_; //storage for matrix on device
-  uint_t* params_;                  //storage for additional parameters on device
-  uint_t base_index_;               //start index of state vector 
-  uint_t chunk_bits_;
-  uint_t* cregs_;
-  uint_t num_creg_bits_;
-  int_t conditional_bit_;
-#ifndef AER_THRUST_CUDA
-  uint_t index_offset_;
-#endif
-public:
-  GateFuncBase()
-  {
-    data_ = NULL;
-    base_index_ = 0;
-    cregs_ = NULL;
-    num_creg_bits_ = 0;
-    conditional_bit_ = -1;
-#ifndef AER_THRUST_CUDA
-    index_offset_ = 0;
-#endif
-  }
-  virtual void set_data(thrust::complex<data_t>* p)
-  {
-    data_ = p;
-  }
-  void set_matrix(thrust::complex<double>* mat)
-  {
-    matrix_ = mat;
-  }
-  void set_params(uint_t* p)
-  {
-    params_ = p;
-  }
-  void set_chunk_bits(uint_t bits)
-  {
-    chunk_bits_ = bits;
-  }
-
-  void set_base_index(uint_t i)
-  {
-    base_index_ = i;
-  }
-  void set_cregs_(uint_t* cbits,uint_t nreg)
-  {
-    cregs_ = cbits;
-    num_creg_bits_ = nreg;
-  }
-  void set_conditional(int_t bit)
-  {
-    conditional_bit_ = bit;
-  }
-
-#ifndef AER_THRUST_CUDA
-  void set_index_offset(uint_t i)
-  {
-    index_offset_ = i;
-  }
-#endif
-
-  __host__ __device__ thrust::complex<data_t>* data(void)
-  {
-    return data_;
-  }
-
-  virtual bool is_diagonal(void)
-  {
-    return false;
-  }
-  virtual int qubits_count(void)
-  {
-    return 1;
-  }
-  virtual int num_control_bits(void)
-  {
-    return 0;
-  }
-  virtual int control_mask(void)
-  {
-    return 1;
-  }
-  virtual bool use_cache(void)
-  {
-    return false;
-  }
-  virtual bool batch_enable(void)
-  {
-    return true;
-  }
-
-  virtual const char* name(void)
-  {
-    return "base function";
-  }
-  virtual uint_t size(int num_qubits)
-  {
-    if(is_diagonal()){
-      chunk_bits_ = num_qubits;
-      return (1ull << num_qubits);
-    }
-    else{
-      chunk_bits_ = num_qubits - (qubits_count() - num_control_bits());
-      return (1ull << (num_qubits - (qubits_count() - num_control_bits())));
-    }
-  }
-
-  virtual __host__ __device__ uint_t thread_to_index(uint_t _tid) const
-  {
-    return _tid;
-  }
-  virtual __host__ __device__ void run_with_cache(uint_t _tid,uint_t _idx,thrust::complex<data_t>* _cache) const
-  {
-    //implemente this in the kernel class
-  }
-  virtual __host__ __device__ double run_with_cache_sum(uint_t _tid,uint_t _idx,thrust::complex<data_t>* _cache) const
-  {
-    //implemente this in the kernel class
-    return 0.0;
-  }
-
-  virtual __host__ __device__ bool check_conditional(uint_t i) const
-  {
-    if(conditional_bit_ < 0)
-      return true;
-
-    uint_t iChunk = i >> chunk_bits_;
-    uint_t n64,i64,ibit;
-    n64 = (num_creg_bits_ + 63) >> 6;
-    i64 = conditional_bit_ >> 6;
-    ibit = conditional_bit_ & 63;
-    return (((cregs_[iChunk*n64 + i64] >> ibit) & 1) != 0);
-  }
-};
-
-//========================================
-  //  gate functions with cache
-//========================================
-template <typename data_t>
-class GateFuncWithCache : public GateFuncBase<data_t>
-{
-protected:
-  int nqubits_;
-public:
-  GateFuncWithCache(uint_t nq)
-  {
-    nqubits_ = nq;
-  }
-
-  bool use_cache(void)
-  {
-    return true;
-  }
-
-    __host__ __device__ virtual uint_t thread_to_index(uint_t _tid) const
-  {
-    uint_t idx,ii,t,j;
-    uint_t* qubits;
-    uint_t* qubits_sorted;
-
-    qubits_sorted = this->params_;
-    qubits = qubits_sorted + nqubits_;
-
-    idx = 0;
-    ii = _tid >> nqubits_;
-    for(j=0;j<nqubits_;j++){
-      t = ii & ((1ull << qubits_sorted[j]) - 1);
-      idx += t;
-      ii = (ii - t) << 1;
-
-      if(((_tid >> j) & 1) != 0){
-        idx += (1ull << qubits[j]);
-      }
-    }
-    idx += ii;
-    return idx;
-  }
-
-  __host__ __device__ void sync_threads() const
-  {
-#ifdef CUDA_ARCH
-    __syncthreads();
-#endif
-  }
-
-  __host__ __device__ void operator()(const uint_t &i) const
-  {
-    if(!this->check_conditional(i))
-      return;
-
-    thrust::complex<data_t> cache[1024];
-    uint_t j,idx;
-    uint_t matSize = 1ull << nqubits_;
-
-    //load data to cache
-    for(j=0;j<matSize;j++){
-      idx = thread_to_index((i << nqubits_) + j);
-      cache[j] = this->data_[idx];
-    }
-
-    //execute using cache
-    for(j=0;j<matSize;j++){
-      idx = thread_to_index((i << nqubits_) + j);
-      this->run_with_cache(j,idx,cache);
-    }
-  }
-
-  virtual int qubits_count(void)
-  {
-    return nqubits_;
-  }
-};
-
-template <typename data_t>
-class GateFuncSumWithCache : public GateFuncBase<data_t>
-{
-protected:
-  int nqubits_;
-public:
-  GateFuncSumWithCache(uint_t nq)
-  {
-    nqubits_ = nq;
-  }
-
-  bool use_cache(void)
-  {
-    return true;
-  }
-
-
-  __host__ __device__ virtual uint_t thread_to_index(uint_t _tid) const
-  {
-    uint_t idx,ii,t,j;
-    uint_t* qubits;
-    uint_t* qubits_sorted;
-
-    qubits_sorted = this->params_;
-    qubits = qubits_sorted + nqubits_;
-
-    idx = 0;
-    ii = _tid >> nqubits_;
-    for(j=0;j<nqubits_;j++){
-      t = ii & ((1ull << qubits_sorted[j]) - 1);
-      idx += t;
-      ii = (ii - t) << 1;
-
-      if(((_tid >> j) & 1) != 0){
-        idx += (1ull << qubits[j]);
-      }
-    }
-    idx += ii;
-    return idx;
-  }
-
-  __host__ __device__ double operator()(const uint_t &i) const
-  {
-    if(!this->check_conditional(i))
-      return 0.0;
-
-    thrust::complex<data_t> cache[1024];
-    uint_t j,idx;
-    uint_t matSize = 1ull << nqubits_;
-    double sum = 0.0;
-
-    //load data to cache
-    for(j=0;j<matSize;j++){
-      idx = thread_to_index((i << nqubits_) + j);
-      cache[j] = this->data_[idx];
-    }
-
-    //execute using cache
-    for(j=0;j<matSize;j++){
-      idx = thread_to_index((i << nqubits_) + j);
-      sum += this->run_with_cache_sum(j,idx,cache);
-    }
-    return sum;
-  }
-
-  virtual int qubits_count(void)
-  {
-    return nqubits_;
-  }
-
-};
-
-//stridded iterator to access diagonal probabilities
-template <typename Iterator>
-class strided_range
-{
-  public:
-
-  typedef typename thrust::iterator_difference<Iterator>::type difference_type;
-
-  struct stride_functor : public thrust::unary_function<difference_type,difference_type>
-  {
-    difference_type stride;
-
-    stride_functor(difference_type stride)
-        : stride(stride) {}
-
-    __host__ __device__
-    difference_type operator()(const difference_type& i) const
-    {
-      if(stride == 1) //statevector
-        return i;
-
-      //density matrix
-      difference_type i_chunk;
-      i_chunk = i / (stride - 1);
-      difference_type ret = stride * i - i_chunk*(stride-1);
-      return ret;
-    }
-  };
-
-  typedef typename thrust::counting_iterator<difference_type>                   CountingIterator;
-  typedef typename thrust::transform_iterator<stride_functor, CountingIterator> TransformIterator;
-  typedef typename thrust::permutation_iterator<Iterator,TransformIterator>     PermutationIterator;
-
-  // type of the strided_range iterator
-  typedef PermutationIterator iterator;
-
-  // construct strided_range for the range [first,last)
-  strided_range(Iterator first, Iterator last, difference_type stride)
-      : first(first), last(last), stride(stride) {}
- 
-  iterator begin(void) const
-  {
-    return PermutationIterator(first, TransformIterator(CountingIterator(0), stride_functor(stride)));
-  }
-
-  iterator end(void) const
-  {
-    if(stride == 1) //statevector
-      return begin() + (last - first);
-
-    //density matrix
-    return begin() + (last - first) / (stride-1);
-  }
-  
-  protected:
-  Iterator first;
-  Iterator last;
-  difference_type stride;
-};
-
-template <typename data_t>
-struct complex_dot_scan : public thrust::unary_function<thrust::complex<data_t>,thrust::complex<data_t>>
-{
-  __host__ __device__
-  thrust::complex<data_t> operator()(thrust::complex<data_t> x) { return thrust::complex<data_t>(x.real()*x.real()+x.imag()*x.imag(),0); }
-};
-
-template <typename data_t>
-struct complex_norm : public thrust::unary_function<thrust::complex<data_t>,thrust::complex<data_t>>
-{
-  __host__ __device__
-  thrust::complex<double> operator()(thrust::complex<data_t> x) { return thrust::complex<double>((double)x.real()*(double)x.real(),(double)x.imag()*(double)x.imag()); }
-};
-
-template<typename data_t>
-struct complex_less
-{
-  typedef thrust::complex<data_t> first_argument_type;
-  typedef thrust::complex<data_t> second_argument_type;
-  typedef bool result_type;
-  __thrust_exec_check_disable__
-    __host__ __device__ bool operator()(const thrust::complex<data_t> &lhs, const thrust::complex<data_t> &rhs) const {return lhs.real() < rhs.real();}
-}; // end less
-
-
-class HostFuncBase
-{
-protected:
-public:
-  HostFuncBase(){}
-
-  virtual void execute(){}
-};
 
 //============================================================================
 // chunk container base class
@@ -474,6 +91,7 @@ protected:
   int_t num_qubits_;                  //total qubits
   int place_id_;                      //index of a container (device index + host)
   uint_t num_chunks_;                 //number of chunks in this container
+  uint_t chunk_index_;                //global chunk index for the first chunk in this container
   uint_t num_buffers_;                //number of buffers (buffer chunks) in this container
   uint_t num_chunk_mapped_;           //number of chunks mapped
   reg_t blocked_qubits_;
@@ -494,6 +112,7 @@ public:
     chunk_bits_ = 0;
     place_id_ = 0;
     num_chunks_ = 0;
+    chunk_index_ = 0;
     num_buffers_ = 0;
     num_chunk_mapped_ = 0;
     conditional_bit_ = -1;
@@ -576,9 +195,14 @@ public:
     keep_conditional_bit_ = keep;
   }
 
+  void set_chunk_index(uint_t chunk_index)
+  {
+    chunk_index_ = chunk_index;
+  }
+
   virtual thrust::complex<data_t>& operator[](uint_t i) = 0;
 
-  virtual uint_t Allocate(int idev,int chunk_bits,int num_qubits,uint_t chunks,uint_t buffers = AER_MAX_BUFFERS,bool multi_shots = false,int matrix_bit = AER_DEFAULT_MATRIX_BITS, bool enable_cuStatevec = false) = 0;
+  virtual uint_t Allocate(int idev,int chunk_bits,int num_qubits,uint_t chunks,uint_t buffers = AER_MAX_BUFFERS,bool multi_shots = false,int matrix_bit = AER_DEFAULT_MATRIX_BITS) = 0;
   virtual void Deallocate(void) = 0;
 
   virtual void Set(uint_t i,const thrust::complex<data_t>& t) = 0;
@@ -607,7 +231,8 @@ public:
   void ExecuteSum2(double* pSum,Function func,uint_t iChunk,uint_t count) const;
 
   virtual reg_t sample_measure(uint_t iChunk,const std::vector<double> &rnds, uint_t stride = 1, bool dot = true,uint_t count = 1) const = 0;
-  virtual thrust::complex<double> norm(uint_t iChunk,uint_t count,uint_t stride = 1,bool dot = true) const = 0;
+  virtual double norm(uint_t iChunk,uint_t count) const;
+  virtual double trace(uint_t iChunk,uint_t row,uint_t count) const;
 
 
   size_t size_of_complex(void)
@@ -690,9 +315,32 @@ public:
   }
   virtual void request_creg_update(void){}
 
-  //apply matrix using cuStatevec
-  virtual void apply_matrix(const uint_t iChunk,const reg_t& qubits,const int_t control_bits,const cvector_t<double> &mat,const uint_t count){}
-  virtual void apply_diagonal_matrix(const uint_t iChunk,const reg_t& qubits,const int_t control_bits,const cvector_t<double> &diag,const uint_t count){}
+  //apply matrix 
+  virtual void apply_matrix(const uint_t iChunk,const reg_t& qubits,const int_t control_bits,const cvector_t<double> &mat,const uint_t count);
+
+  //apply diagonal matrix
+  virtual void apply_diagonal_matrix(const uint_t iChunk,const reg_t& qubits,const int_t control_bits,const cvector_t<double> &diag,const uint_t count);
+
+  //apply (controlled) X
+  virtual void apply_X(const uint_t iChunk,const reg_t& qubits,const uint_t count);
+
+  //apply (controlled) Y
+  virtual void apply_Y(const uint_t iChunk,const reg_t& qubits,const uint_t count);
+
+  //apply (controlled) phase
+  virtual void apply_phase(const uint_t iChunk,const reg_t& qubits,const int_t control_bits,const std::complex<double> phase,const uint_t count);
+
+  //apply (controlled) swap gate
+  virtual void apply_swap(const uint_t iChunk,const reg_t& qubits,const int_t control_bits,const uint_t count);
+
+  //apply permutation
+  virtual void apply_permutation(const uint_t iChunk,const reg_t& qubits,const std::vector<std::pair<uint_t, uint_t>> &pairs, const uint_t count);
+
+  //get probabilities of chunk
+  virtual void probabilities(std::vector<double>& probs, const uint_t iChunk, const reg_t& qubits) const;
+
+  //Pauli expectation values
+  virtual double expval_pauli(const uint_t iChunk,const reg_t& qubits,const std::string &pauli,const complex_t initial_phase) const;
 
 protected:
   int convert_blocked_qubit(int qubit)
@@ -775,6 +423,7 @@ void ChunkContainer<data_t>::Execute(Function func,uint_t iChunk,uint_t count)
 {
   set_device();
 
+  func.set_base_index((chunk_index_ + iChunk) << chunk_bits_);
   func.set_data( chunk_pointer(iChunk) );
   func.set_matrix( matrix_pointer(iChunk) );
   func.set_params( param_pointer(iChunk) );
@@ -845,6 +494,7 @@ void ChunkContainer<data_t>::ExecuteSum(double* pSum,Function func,uint_t iChunk
 
   set_device();
 
+  func.set_base_index((chunk_index_ + iChunk) << chunk_bits_);
   func.set_data( chunk_pointer(iChunk) );
   func.set_matrix( matrix_pointer(iChunk) );
   func.set_params( param_pointer(iChunk) );
@@ -1010,6 +660,7 @@ void ChunkContainer<data_t>::ExecuteSum2(double* pSum,Function func,uint_t iChun
 
   set_device();
 
+  func.set_base_index((chunk_index_ + iChunk) << chunk_bits_);
   func.set_data( chunk_pointer(iChunk) );
   func.set_matrix( matrix_pointer(iChunk) );
   func.set_params( param_pointer(iChunk) );
@@ -1168,6 +819,171 @@ void ChunkContainer<data_t>::update_pow2_qubits(void)
     n >>= 1;
   }
 }
+
+template <typename data_t>
+void ChunkContainer<data_t>::apply_matrix(const uint_t iChunk,const reg_t& qubits,const int_t control_bits,const cvector_t<double> &mat,const uint_t count)
+{
+  const size_t N = qubits.size() - control_bits;
+  auto qubits_sorted = qubits;
+  std::sort(qubits_sorted.begin(), qubits_sorted.end());
+
+  if(N == 1){
+    if(control_bits == 0)
+      Execute(MatrixMult2x2<data_t>(mat,qubits[0]), iChunk, count);
+    else  //2x2 matrix with control bits
+      Execute(MatrixMult2x2Controlled<data_t>(mat,qubits), iChunk, count);
+  }
+  else if(N == 2){
+    Execute(MatrixMult4x4<data_t>(mat,qubits[0],qubits[1]), iChunk, count);
+  }
+  else{
+    if(N <= 10){
+      int i;
+      for(i=0;i<N;i++){
+        qubits_sorted.push_back(qubits[i]);
+      }
+      StoreMatrix(mat, iChunk);
+      StoreUintParams(qubits_sorted, iChunk);
+
+      Execute(MatrixMultNxN<data_t>(N), iChunk, count);
+    }
+    else{
+      cvector_t<double> matLU;
+      reg_t params;
+      MatrixMultNxN_LU<data_t> f(mat,qubits_sorted,matLU,params);
+
+      StoreMatrix(matLU, iChunk);
+      StoreUintParams(params, iChunk);
+
+      Execute(f, iChunk, count);
+    }
+  }
+}
+
+template <typename data_t>
+void ChunkContainer<data_t>::apply_diagonal_matrix(const uint_t iChunk,const reg_t& qubits,const int_t control_bits,const cvector_t<double> &diag,const uint_t count)
+{
+  const size_t N = qubits.size() - control_bits;
+
+  if(N == 1){
+    if(control_bits == 0)
+      Execute(DiagonalMult2x2<data_t>(diag,qubits[0]), iChunk, count);
+    else
+      Execute(DiagonalMult2x2Controlled<data_t>(diag,qubits), iChunk, count);
+  }
+  else if(N == 2){
+    Execute(DiagonalMult4x4<data_t>(diag,qubits[0],qubits[1]), iChunk, count);
+  }
+  else{
+    StoreMatrix(diag, iChunk);
+    StoreUintParams(qubits, iChunk);
+
+    Execute(DiagonalMultNxN<data_t>(qubits), iChunk, count);
+  }
+}
+
+template <typename data_t>
+void ChunkContainer<data_t>::apply_X(const uint_t iChunk,const reg_t& qubits,const uint_t count)
+{
+  Execute(CX_func<data_t>(qubits), iChunk, count);
+}
+
+template <typename data_t>
+void ChunkContainer<data_t>::apply_Y(const uint_t iChunk,const reg_t& qubits,const uint_t count)
+{
+  Execute(CY_func<data_t>(qubits), iChunk, count);
+}
+
+template <typename data_t>
+void ChunkContainer<data_t>::apply_phase(const uint_t iChunk,const reg_t& qubits,const int_t control_bits,const std::complex<double> phase,const uint_t count)
+{
+  Execute(phase_func<data_t>(qubits,*(thrust::complex<double>*)&phase), iChunk, count );
+}
+
+template <typename data_t>
+void ChunkContainer<data_t>::apply_swap(const uint_t iChunk,const reg_t& qubits,const int_t control_bits,const uint_t count)
+{
+  Execute(CSwap_func<data_t>(qubits), iChunk, count);
+}
+
+template <typename data_t>
+void ChunkContainer<data_t>::apply_permutation(const uint_t iChunk,const reg_t& qubits,const std::vector<std::pair<uint_t, uint_t>> &pairs, const uint_t count)
+{
+  const size_t N = qubits.size();
+  auto qubits_sorted = qubits;
+  std::sort(qubits_sorted.begin(), qubits_sorted.end());
+
+  reg_t params;
+  Permutation<data_t> f(qubits_sorted,qubits,pairs,params);
+
+  StoreUintParams(params, iChunk);
+
+  Execute(f, iChunk, count);
+}
+
+template <typename data_t>
+void ChunkContainer<data_t>::probabilities(std::vector<double>& probs, const uint_t iChunk, const reg_t& qubits) const
+{
+  const size_t N = qubits.size();
+  const int_t DIM = 1 << N;
+  probs.resize(DIM);
+
+  if(N == 1){ //special case for 1 qubit (optimized for measure)
+    ExecuteSum2(&probs[0],probability_1qubit_func<data_t>(qubits[0]), iChunk, 1);
+  }
+  else{
+    for(int_t i=0;i<DIM;i++){
+      ExecuteSum(&probs[i],probability_func<data_t>(qubits,i), iChunk, 1);
+    }
+  }
+}
+
+template <typename data_t>
+double ChunkContainer<data_t>::norm(uint_t iChunk,uint_t count) const
+{
+  double ret;
+  ExecuteSum(&ret,norm_func<data_t>(), iChunk, count);
+
+  return ret;
+}
+
+template <typename data_t>
+double ChunkContainer<data_t>::trace(uint_t iChunk,uint_t row,uint_t count) const
+{
+  double ret;
+  ExecuteSum(&ret,trace_func<data_t>(row), iChunk, count);
+
+  return ret;
+}
+
+template <typename data_t>
+double ChunkContainer<data_t>::expval_pauli(const uint_t iChunk,const reg_t& qubits,const std::string &pauli,const complex_t initial_phase) const
+{
+  uint_t x_mask, z_mask, num_y, x_max;
+  std::tie(x_mask, z_mask, num_y, x_max) = pauli_masks_and_phase(qubits, pauli);
+
+  // Special case for only I Paulis
+  if (x_mask + z_mask == 0) {
+    thrust::complex<double> ret = norm(iChunk, 1);
+    return ret.real() + ret.imag();
+  }
+  double ret;
+  // specialize x_max == 0
+  if(x_mask == 0) {
+    ExecuteSum(&ret, expval_pauli_Z_func<data_t>(z_mask), iChunk,  1 );
+    return ret;
+  }
+
+  // Compute the overall phase of the operator.
+  // This is (-1j) ** number of Y terms modulo 4
+  auto phase = std::complex<data_t>(initial_phase);
+  add_y_phase(num_y, phase);
+  ExecuteSum(&ret, expval_pauli_XYZ_func<data_t>(x_mask, z_mask, x_max, phase), iChunk, 1 );
+  return ret;
+}
+
+
+
 
 //------------------------------------------------------------------------------
 } // end namespace QV
