+++ src/simulators/unitary/unitarymatrix_thrust.hpp
@@ -280,54 +280,13 @@ void UnitaryMatrixThrust<data_t>::set_num_qubits(size_t num_qubits) {
   BaseVector::set_num_qubits(2 * num_qubits);
 }
 
-template <typename data_t>
-class trace_func : public GateFuncBase<data_t>
-{
-protected:
-  uint_t rows_;
-public:
-  trace_func(uint_t nrow)
-  {
-    rows_ = nrow;
-  }
-  bool is_diagonal(void)
-  {
-    return true;
-  }
-  uint_t size(int num_qubits)
-  {
-    this->chunk_bits_ = num_qubits;
-    return rows_;
-  }
-
-  __host__ __device__ double operator()(const uint_t &i) const
-  {
-    thrust::complex<data_t> q;
-    thrust::complex<data_t>* vec;
-
-    uint_t iChunk = (i / rows_);
-    uint_t lid = i - (iChunk * rows_);
-    uint_t idx = (iChunk << this->chunk_bits_) + lid*(rows_ + 1);
-
-    vec = this->data_;
-    q = vec[idx];
-    return q.real();
-  }
-
-  const char* name(void)
-  {
-    return "trace";
-  }
-};
 
 template <class data_t>
 std::complex<double> UnitaryMatrixThrust<data_t>::trace() const 
 {
   thrust::complex<double> sum;
-  double ret;
 
-  BaseVector::apply_function_sum(&ret,trace_func<data_t>(rows_),false);
-  sum = ret;
+  sum = BaseVector::chunk_.trace(rows_, 1);
 
 #ifdef AER_DEBUG
   BaseVector::DebugMsg("trace",sum);
